requirements.txt
streamlit
PIL
google-generativeai
python-dotenv